# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dbaas adapter API",
        "contact": {}
    },
    "basePath": "/api/{apiVersion}/dbaas/adapter",
    "paths": {
        "/physical_database/force_registration": {
            "get": {
                "description": "Force this adapter to immediately register itself in dbaas-aggregator.\nAdapter initiates background task that tries to register physical database in dbaas-aggregator,\nand responds with status 202 before the background task finishes.",
                "tags": [
                    "Common dbaas adapter operations"
                ],
                "summary": "Force physical database registration",
                "parameters": [
                    {
                        "enum": [
                            "v1",
                            "v2"
                        ],
                        "type": "string",
                        "default": "v2",
                        "description": "API version of dbaas adapter",
                        "name": "apiVersion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "if physical database registration process has been started successfully.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}